Move "io.cache" package inside "io.http"
